Mutating request should actually do some mutation
