+++ src/Features/LanguageServer/ProtocolUnitTests/Ordering/NonMutatingRequestHandler.cs
@@ -4,21 +4,44 @@
 
 using System;
 using System.Composition;
+using System.Threading;
+using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.CodeAnalysis.LanguageServer.Handler;
 
 namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.RequestOrdering
 {
     [Shared, ExportLspMethod(MethodName, mutatesSolutionState: false)]
-    internal class NonMutatingRequestHandler : AbstractTestRequestHandler
+    internal class NonMutatingRequestHandler : AbstractRequestHandler<OrderedLspRequest, OrderedLspResponse>
     {
         public const string MethodName = nameof(NonMutatingRequestHandler);
+        private const int Delay = 100;
 
         [ImportingConstructor]
         [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
         public NonMutatingRequestHandler(ILspSolutionProvider solutionProvider)
             : base(solutionProvider)
         {
+        }
+
+        public override async Task<OrderedLspResponse> HandleRequestAsync(OrderedLspRequest request, RequestContext context, CancellationToken cancellationToken)
+        {
+            var response = new OrderedLspResponse();
+
+            response.Solution = context.Solution;
+            response.RequestOrder = request.RequestOrder;
+            response.StartTime = DateTime.UtcNow;
+
+            await Task.Delay(Delay, cancellationToken).ConfigureAwait(false);
+
+            // some busy work
+            response.ToString();
+
+            await Task.Delay(Delay, cancellationToken).ConfigureAwait(false);
+
+            response.EndTime = DateTime.UtcNow;
+
+            return response;
         }
     }
 }
