+++ src/Features/LanguageServer/ProtocolUnitTests/Ordering/OrderedLspResponse.cs
@@ -11,5 +11,7 @@ namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.RequestOrdering
         public int RequestOrder { get; set; }
         public DateTime StartTime { get; set; }
         public DateTime EndTime { get; set; }
+
+        public Solution Solution { get; set; }
     }
 }
